+++ analytics/src/main/java/com/agileburo/anytype/analytics/props/Props.kt
@@ -1,0 +1,9 @@
+package com.agileburo.anytype.analytics.props
+
+data class Props(val map: Map<String?, Any?>) {
+    private val default = map.withDefault { null }
+
+    companion object {
+        fun empty() = Props(emptyMap())
+    }
+}