+++ analytics/src/main/java/com/agileburo/anytype/analytics/props/UserProperty.kt
@@ -1,0 +1,3 @@
+package com.agileburo.anytype.analytics.props
+
+interface UserProperty