+++ analytics/src/main/java/com/agileburo/anytype/analytics/tracker/Tracker.kt
@@ -1,0 +1,3 @@
+package com.agileburo.anytype.analytics.tracker
+
+interface Tracker