+++ app/src/main/java/com/agileburo/anytype/app/AndroidApplication.kt
@@ -6,13 +6,19 @@ import androidx.emoji.text.EmojiCompat
 import androidx.emoji.text.FontRequestEmojiCompatConfig
 import com.agileburo.anytype.BuildConfig
 import com.agileburo.anytype.R
+import com.agileburo.anytype.analytics.tracker.AmplitudeTracker
 import com.agileburo.anytype.core_utils.tools.CrashlyticsTree
 import com.agileburo.anytype.di.common.ComponentManager
 import com.agileburo.anytype.di.main.*
+import com.amplitude.api.Amplitude
 import com.facebook.stetho.Stetho
 import timber.log.Timber
+import javax.inject.Inject
 
 class AndroidApplication : Application() {
+
+    @Inject
+    lateinit var amplitudeTracker: AmplitudeTracker
 
     private val main: MainComponent by lazy {
         DaggerMainComponent
@@ -22,6 +28,7 @@ class AndroidApplication : Application() {
             .configModule(ConfigModule())
             .utilModule(UtilModule())
             .deviceModule(DeviceModule())
+            .analyticsModule(AnalyticsModule())
             .build()
     }
 
@@ -31,7 +38,8 @@ class AndroidApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-
+        main.inject(this)
+        setupAnalytics()
         setupEmojiCompat()
         setupTimber()
         setupStetho()
@@ -58,5 +66,9 @@ class AndroidApplication : Application() {
     private fun setupStetho() {
         if (BuildConfig.DEBUG)
             Stetho.initializeWithDefaults(this)
+    }
+
+    private fun setupAnalytics() {
+        Amplitude.getInstance().initialize(this, getString(R.string.amplitude_api_key))
     }
 }