+++ presentation/src/main/java/com/agileburo/anytype/presentation/splash/SplashViewModelFactory.kt
@@ -2,6 +2,7 @@ package com.agileburo.anytype.presentation.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.agileburo.anytype.analytics.base.Analytics
 import com.agileburo.anytype.domain.auth.interactor.CheckAuthorizationStatus
 import com.agileburo.anytype.domain.auth.interactor.LaunchAccount
 import com.agileburo.anytype.domain.auth.interactor.LaunchWallet
@@ -14,7 +15,8 @@ import com.agileburo.anytype.domain.auth.interactor.LaunchWallet
 class SplashViewModelFactory(
     private val checkAuthorizationStatus: CheckAuthorizationStatus,
     private val launchAccount: LaunchAccount,
-    private val launchWallet: LaunchWallet
+    private val launchWallet: LaunchWallet,
+    private val analytics: Analytics
 ) : ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
@@ -22,6 +24,7 @@ class SplashViewModelFactory(
         SplashViewModel(
             checkAuthorizationStatus = checkAuthorizationStatus,
             launchAccount = launchAccount,
-            launchWallet = launchWallet
+            launchWallet = launchWallet,
+            analytics = analytics
         ) as T
 }